# Simplify string reply check