The go compiler is smart enough to optimize string([]byte) conversion.

```
name       old time/op    new time/op    delta
DoPing-12    29.2µs ± 3%    29.3µs ± 4%   ~     (p=0.730 n=9+9)

name       old alloc/op   new alloc/op   delta
DoPing-12     0.00B          0.00B        ~     (all equal)

name       old allocs/op  new allocs/op  delta
DoPing-12      0.00           0.00        ~     (all equal)
```